requirement.txt
flask
flask-mysql
flask-restful
PyJWT